% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorpusFromBibData.R
\name{CorpusFromBibData}
\alias{CorpusFromBibData}
\title{Corpus from an article database}
\usage{
CorpusFromBibData(bibData, bibUnits, controlList, stopwords, wordsToRemove,
  replaceWords)
}
\arguments{
\item{bibData}{a vector containing the key terms to search}

\item{bibUnits}{a vector containing bibliometric units of analysis (e.g., Title, Abstract, Keywords, Journal, Authors, Year)}

\item{controlList}{a vector of transformations that will be applied to the corpus}

\item{stopwords}{a vector of stopwords to be removed from the corpus}

\item{wordsToRemove}{a vector of custom words to be removed}

\item{replaceWords}{a data frame of custom words and its corresponding replacement word}
}
\value{
a corpus object
}
\description{
Obtaining of a corpus from an article database
}
\examples{


}

