% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/First.R
\name{First}
\alias{First}
\title{First element}
\usage{
First(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the first element of the incoming object
}
\description{
Returns the first element of an array
}
\examples{


}

