% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHREF.R
\name{GetHREF}
\alias{GetHREF}
\title{Get HREF attribute}
\usage{
GetHREF(nodeSet)
}
\arguments{
\item{nodeSet}{a html node or node set}
}
\value{
the href attribute from the node (or the nodes)
}
\description{
Get the href attribute from a html object
}
\examples{


}

