% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRISElement.R
\name{GetRISElement}
\alias{GetRISElement}
\title{Get text from a RIS element}
\usage{
GetRISElement(x, pattern, replacement = "", collapse = ";")
}
\arguments{
\item{x}{a vector}

\item{pattern}{the pattern to match the line contents}

\item{replacement}{the text for replace the matched pattern (empty by default)}

\item{collapse}{the symbol for collapse the resulting array of contents in the text line}
}
\value{
the text line of a RIS element
}
\description{
Get the text line from a RIS element (used internally for the GetRISList function)
}
\examples{


}

