% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRISList.R
\name{GetRISList}
\alias{GetRISList}
\title{Get text from a RIS bibliometric unit}
\usage{
GetRISList(data, pattern, replacement = "", collapse = ";")
}
\arguments{
\item{data}{a list containing the RIS info of each article}

\item{pattern}{the pattern to match the line contents}

\item{replacement}{the text for replace the matched pattern (empty by default)}

\item{collapse}{the symbol for collapse the resulting array of contents in the text line}
}
\value{
the text lines of a RIS bibliometric unit
}
\description{
Get the text lines from a RIS bibliometric unit
}
\examples{


}

