  \name{KDCorpus}
  \docType{data}
  \alias{KDCorpus}
  \title{Corpus of articles and keywords related to the topic of Knowledge domain visualization}
  \description{
    This corpus is composed by keywords inside articles related with Knowledge Domain Visualization field.
  }
  \usage{KDCorpus}
  \format{A corpus containing the information of terms associated to 757 articles related to the keywords \code{knowledge domain visualization}, \code{document clustering}, \code{co-words} and \code{co-citation}.}
  \keyword{datasets}