\name{KDDTM}
\docType{data}
\alias{KDDTM}
\title{Document-term matrix of articles and keywords related to the topic of Knowledge domain visualization}
\description{
  This document-term matrix was obtained from a corpus composed by keywords inside articles related with Knowledge Domain Visualization field.
}
\usage{KDDTM}
\format{A document-term matrix of 2486 terms associated to 555 articles related to the keywords \code{knowledge domain visualization}, \code{document clustering}, \code{co-words} and \code{co-citation}.}
\keyword{datasets}