\name{KDVizData}
\docType{data}
\alias{KDVizData}
\title{Data about the topic of Knowledge domain visualization}
\description{
  This data set is bibliographic information retrieved with \code{\link{ArticleSearch}} funtion from this package, it contains the main information of articles related with Knowledge Domain Visualization field.
}
\usage{KDVizData}
\format{A dataframe containing the information of \code{Keywords} and \code{URLs} from 757 articles related to the keywords \code{knowledge domain visualization}, \code{document clustering}, \code{co-words} and \code{co-citation}.}
\keyword{datasets}