% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Last.R
\name{Last}
\alias{Last}
\title{Last element}
\usage{
Last(x)
}
\arguments{
\item{x}{a vector}
}
\value{
the last element of the incoming object
}
\description{
Returns the last element of an array
}
\examples{


}

