% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer.mpa.corpus.R
\name{leer.mpa.corpus}
\alias{leer.mpa.corpus}
\title{Reads a corpus and passes it to mpa format}
\usage{
leer.mpa.corpus(corpus)
}
\arguments{
\item{corpus}{a corpus object}
}
\value{
a vector containing the term list per document
}
\description{
Returns the content of a corpus object to use mpa package methods
}
\examples{


}

