% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriz.mpa.corpus.R
\name{matriz.mpa.corpus}
\alias{matriz.mpa.corpus}
\title{Calculation of co-occurrences matrix and matrix associations from a corpus}
\usage{
matriz.mpa.corpus(corpus, fmin = 3, cmin = 3)
}
\arguments{
\item{corpus}{a corpus object}

\item{fmin}{minimal appearance frequency of key words inside the corpus}

\item{cmin}{minimal co-occurrence between words}
}
\value{
a list that contains the associations and the co-ocurrence matrices, the vector of words and a lexical table (obtained from the original matriz.mpa function0)
}
\description{
Similar to the mpa package, it calculates the co-occurrences matrix and the matrix associations from the resulting object of the leer.mpa.corpus function
}
\examples{


}

