% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScienceDirectArticles.R
\name{ScienceDirectArticles}
\alias{ScienceDirectArticles}
\title{Article search from ScienceDirect}
\usage{
ScienceDirectArticles(keywords, size, addInfo = FALSE, infoList)
}
\arguments{
\item{keywords}{a vector containing the key terms to search}

\item{size}{the number of articles from which the information is extracted}

\item{addInfo}{a logical value indicating whether the info of abstract, keyword, journalTitle, journalVol and authorName should be retrieved by each article}

\item{infoList}{a data frame of titles and URLs to skip the first step of getting the main information of each article}
}
\value{
a data frame containing the information requested in the function call
}
\description{
Search articles from the ScienceDirect database by specifying a list of key terms
}
\examples{


}

