% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseRate.R
\name{SparseRate}
\alias{SparseRate}
\title{Sparse rate to remove a proportion of terms}
\usage{
SparseRate(termFreq, dtm)
}
\arguments{
\item{termFreq}{a minimum allowed frecuency to define the sparse of the terms}

\item{dtm}{document-term matrix}
}
\value{
sparse percentage of non empty elements
}
\description{
Returns the threshold to decide the sparse rate to use depending on the minimum allowed frequency of the terms in the document term matrix
}
\examples{


}

