% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordGroupDTM.R
\name{WordGroupDTM}
\alias{WordGroupDTM}
\title{Document-term matrix and SCA by word group}
\usage{
WordGroupDTM(class, dtm, wordClasses, graph = FALSE)
}
\arguments{
\item{class}{the number of the class to be partitioned}

\item{dtm}{a document-term matrix}

\item{wordClasses}{the value resulting from performing a KDSummary}

\item{graph}{a logical value. If TRUE, the knowledge domain map of the corresponding class is plotted}
}
\value{
the doc-term matrix and the SCA object from the document and word group
}
\description{
A portion of an entire document-term matrix depending on the class found using the kdSummary function
}
\examples{


}

