% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WordsInCorpus.R
\name{WordsInCorpus}
\alias{WordsInCorpus}
\title{See words inside a corpus}
\usage{
WordsInCorpus(corpus)
}
\arguments{
\item{corpus}{a corpus object}
}
\value{
a vector of words
}
\description{
A function to return all words found in a corpus
}
\examples{


}

