% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMFromCorpus.R
\name{DTMFromCorpus}
\alias{DTMFromCorpus}
\title{Document-term matrix from a corpus}
\usage{
DTMFromCorpus(corpus, rowNames)
}
\arguments{
\item{corpus}{a corpus object}

\item{rowNames}{the row names of the matrix where the corpus comes from}
}
\value{
a document-term matrix
}
\description{
Obtaining of a binary document-term matrix from a corpus removing null rows
}
\examples{


}

