% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDSummary.R
\name{KDSummary}
\alias{KDSummary}
\title{Summary of word groups}
\usage{
KDSummary(matriz.mpa, mpa)
}
\arguments{
\item{matriz.mpa}{vector from the different words that appears in the corpus (returned value by matriz.mpa function)}

\item{mpa}{a list of values resulting from the mpa function}
}
\value{
a list of objects to summarize the term clustering mpa method
}
\description{
Summary of word groups used to visualize knowledge domains
}
\examples{


}

