% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRIS.R
\name{ReadRIS}
\alias{ReadRIS}
\title{Read a RIS file}
\usage{
ReadRIS(risFile, fileName, saveRda = FALSE, saveCSV = FALSE)
}
\arguments{
\item{risFile}{a file of RIS extension}

\item{fileName}{a character giving the name of the resulting file to export}

\item{saveRda}{a logical value that indicates whether the file should be saved or not in a Rda file}

\item{saveCSV}{a logical value that indicates whether the file should be saved or not in a csv file}
}
\value{
a data frame of bibliometric units by each article
}
\description{
Read the entire info from a RIS file
}
\examples{


}

