% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceByList.R
\name{ReplaceByList}
\alias{ReplaceByList}
\title{Replace a list of words by a pair list}
\usage{
ReplaceByList(corpus, wordsFile)
}
\arguments{
\item{corpus}{a corpus object}

\item{wordsFile}{a file with custom words to be replaced (first column with the replacement words, second column with the original words; separated by tabulation)}
}
\value{
a corpus with replaced words
}
\description{
A process similar to lemmatization with a custom dictionary file in the form of a data frame of custom words and its corresponding replacement word
}
\examples{


}
\author{
Camila Góngora \email{mcgongoraa@unal.edu.co},
Andrés Palacios \email{anfpalacioscl@unal.edu.co}
}

