% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFreqByWord.R
\name{TermFreqByWord}
\alias{TermFreqByWord}
\title{Term frequency by a specific words}
\usage{
TermFreqByWord(termFreqTable, word)
}
\arguments{
\item{termFreqTable}{a TermFrequency table}

\item{word}{a custom word to be matched}
}
\value{
a list of terms and their respective frecuencies
}
\description{
Returns the frequency of the terms containing a specific word
}
\examples{


}

