% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermFrequency.R
\name{TermFrequency}
\alias{TermFrequency}
\title{Term frequency}
\usage{
TermFrequency(x)
}
\arguments{
\item{x}{a corpus or a document-term matrix object}
}
\value{
a list of terms and their respective frecuencies
}
\description{
A list of terms and their absolute frequencies in a corpus or a document-term matrix
}
\examples{


}

